+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Add `start()` and `end()` method to the `Region` trait.
 - Much faster `OverlapIterator`.
+- Add `BlockDevice` trait, along with `BlockIdx` and `BlockCount` types.
 
 ## [0.3.1] - 2023-12-04